Rework classes to streamline inclusion of new datasets from the collections
