+++ torchhd/datasets/__init__.py
@@ -6,7 +6,12 @@ from torchhd.datasets.airfoil_self_noise import AirfoilSelfNoise
 from torchhd.datasets.emg_hand_gestures import EMGHandGestures
 from torchhd.datasets.pamap import PAMAP
 from torchhd.datasets.ccpp import CyclePowerPlant
+from torchhd.datasets.collection_datasets import CollectionDataset
+from torchhd.datasets.collection_datasets import DatasetFourFold
+from torchhd.datasets.collection_datasets import DatasetTrainTest
 from torchhd.datasets.abalone import Abalone
+from torchhd.datasets.adult import Adult
+
 
 __all__ = [
     "BeijingAirQuality",
@@ -17,5 +22,9 @@ __all__ = [
     "EMGHandGestures",
     "PAMAP",
     "CyclePowerPlant",
+    "CollectionDataset",
+    "DatasetFourFold",
+    "DatasetTrainTest",
     "Abalone",
+    "Adult",
 ]
